feat(asset-selector): add search hook and dropdown
